Pin dependencies and fix flake8 errors to fix tests and fresh installs (#224)
* Pin dependencies to fix tests

psycopg2 had to be pinned to a lower version because of pinning in
pgcli

Old redis pinning conflicted with new celery versions. As a result,
celery was bumped (tests seemed not to care) and celery was bumped
accordingly to 4.3.0 or greater. Alternatively, celery could have been
pinned to >=4.2.0,<4.3.0.

* Fix flake8 errors

* Revert redis update, pin celery to below 4.3.0

Some tests broke with new redis